+++ packages/thirdweb/src/exports/transaction.ts
@@ -25,6 +25,10 @@ export {
   estimateGas,
   type EstimateGasResult,
 } from "../transaction/actions/estimate-gas.js";
+export {
+  estimateGasCost,
+  type EstimateGasCostResult,
+} from "../transaction/actions/estimate-gas-cost.js";
 export { waitForReceipt } from "../transaction/actions/wait-for-tx-receipt.js";
 export {
   sendTransaction,